+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 - added option to disable questions and notifications (Thorsten)
 - added new mobile first, touch-friendly default layout (Thorsten)
 - added HTTPS support for local Docker development (Thorsten)
+- added Symfony HttpFoundation (Thorsten)
 - added Monolog v3 as logging solution (Thorsten)
 - removed experimental large permissions support with sections (Thorsten)
 - removed broken link verification (Thorsten)